catalog: Support for integer and float types in variables and annotations. NUMERIC type has been removed. Closes #545.

+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb/MongoDBUtils.java
@@ -663,7 +663,7 @@ class MongoDBUtils {
             if (variableType == Variable.VariableType.BOOLEAN) {
                 type = QueryParam.Type.BOOLEAN;
 
-            } else if (variableType == Variable.VariableType.NUMERIC) {
+            } else if (variableType == Variable.VariableType.DOUBLE) {
                 type = QueryParam.Type.DECIMAL;
             }
         }
@@ -729,7 +729,7 @@ class MongoDBUtils {
                 if (variableType == Variable.VariableType.BOOLEAN) {
                     type = AbstractDBAdaptor.FilterOption.Type.BOOLEAN;
 
-                } else if (variableType == Variable.VariableType.NUMERIC) {
+                } else if (variableType == Variable.VariableType.DOUBLE) {
                     type = AbstractDBAdaptor.FilterOption.Type.NUMERICAL;
                 }
             }
